requirements.txt
streamlit>=1.32.0
gspread>=6.0.0
google-auth>=2.28.0
google-api-python-client>=2.120.0
pandas>=2.0.0
requests>=2.31.0
anthropic>=0.25.0
openai>=1.30.0
google-genai>=0.5.0
mistralai>=0.4.0
groq>=0.9.0
